+++ examples/adder/src/lib.rs
@@ -59,7 +59,7 @@ mod tests {
     #[tokio::test]
     async fn embedded_abi_test() -> anyhow::Result<()> {
         let wasm = fs::read("res/adder.wasm").await?;
-        let worker = unc_workspaces::sandbox().await?;
+        let worker = utility_workspaces::sandbox().await?;
         let contract = worker.dev_deploy(&wasm).await?;
 
         let res = contract.view("__contract_abi").await?;