+++ examples/non-fungible-token/tests/workspaces/test_approval.rs
@@ -2,7 +2,7 @@ use crate::utils::init;
 use unc_contract_standards::non_fungible_token::Token;
 
 use unc_sdk::AccountId;
-use unc_workspaces::types::UncToken;
+use utility_workspaces::types::UncToken;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 
@@ -13,7 +13,7 @@ const ONE_YOCTO: UncToken = UncToken::from_attounc(1);
 
 #[tokio::test]
 async fn simulate_simple_approve() -> anyhow::Result<()> {
-    let worker = unc_workspaces::sandbox().await?;
+    let worker = utility_workspaces::sandbox().await?;
     let (nft_contract, alice, token_receiver_contract, _) = init(&worker).await?;
 
     // root approves alice
@@ -103,7 +103,7 @@ async fn simulate_simple_approve() -> anyhow::Result<()> {
 
 #[tokio::test]
 async fn simulate_approval_with_call() -> anyhow::Result<()> {
-    let worker = unc_workspaces::sandbox().await?;
+    let worker = utility_workspaces::sandbox().await?;
     let (nft_contract, _, _, approval_receiver_contract) = init(&worker).await?;
 
     let res = nft_contract
@@ -133,7 +133,7 @@ async fn simulate_approval_with_call() -> anyhow::Result<()> {
 
 #[tokio::test]
 async fn simulate_approved_account_transfers_token() -> anyhow::Result<()> {
-    let worker = unc_workspaces::sandbox().await?;
+    let worker = utility_workspaces::sandbox().await?;
     let (nft_contract, alice, _, _) = init(&worker).await?;
 
     // root approves alice
@@ -166,7 +166,7 @@ async fn simulate_approved_account_transfers_token() -> anyhow::Result<()> {
 
 #[tokio::test]
 async fn simulate_revoke() -> anyhow::Result<()> {
-    let worker = unc_workspaces::sandbox().await?;
+    let worker = utility_workspaces::sandbox().await?;
     let (nft_contract, alice, token_receiver_contract, _) = init(&worker).await?;
 
     // root approves alice
@@ -250,7 +250,7 @@ async fn simulate_revoke() -> anyhow::Result<()> {
 
 #[tokio::test]
 async fn simulate_revoke_all() -> anyhow::Result<()> {
-    let worker = unc_workspaces::sandbox().await?;
+    let worker = utility_workspaces::sandbox().await?;
     let (nft_contract, alice, token_receiver_contract, _) = init(&worker).await?;
 
     // root approves alice